Refactor MessageList to use DBClass for DB operations
Updated MessageList.cshtml to display sender names separately.
Refactored MessageList.cshtml.cs to use DBClass for database operations:
- Added using directive for Lab1_Part3_Johnson_Imlay.Pages.DB.
- Replaced inline SQL connection string with DBClass methods.
- Changed types of Messages, SelectedMessage, and Senders properties.
- Removed private methods and replaced with DBClass method calls.
Added new methods in DBClass.cs for database operations:
- GetMessages, GetMessageSenders, GetMessageByID.
- Added MessageModel and UserModel classes.

Not working properly, but this is the version that I will be working on today.
