+++ Lab1_Part3_Johnson_Imlay/Pages/Admin/Messages/MessageList.cshtml
@@ -15,7 +15,7 @@
                 @foreach (var sender in Model.Senders)
                 {
                     <option value="@sender.UserID" selected="@(Model.SelectedSenderID == sender.UserID ? "selected" : null)">
-                        @sender.FullName
+                        @sender.FirstName @sender.LastName
                     </option>
                 }
             </select>
